refactor(utils): mv waitForCondition to @streamr/utils
