+++ packages/broker/test/integration/broker-subscriptions.test.ts
@@ -2,8 +2,8 @@ import { Wallet } from '@ethersproject/wallet'
 import mqtt, { AsyncMqttClient } from 'async-mqtt'
 import StreamrClient, { Stream, StreamPermission } from 'streamr-client'
 import { Tracker } from '@streamr/network-tracker'
-import { fastWallet, fetchPrivateKeyWithGas, waitForCondition } from 'streamr-test-utils'
-import { wait } from '@streamr/utils'
+import { fastWallet, fetchPrivateKeyWithGas } from 'streamr-test-utils'
+import { wait, waitForCondition } from '@streamr/utils'
 import { Broker } from '../../src/broker'
 import { startBroker, createClient, createTestStream, getStreamParts, startTestTracker } from '../utils'
