+++ packages/broker/test/integration/plugins/brubeckMiner/BrubeckMinerPlugin.test.ts
@@ -2,10 +2,10 @@ import { Server } from 'http'
 import { once } from 'events'
 import { Wallet } from 'ethers'
 import express, { Request, Response } from 'express'
-import { Logger, toEthereumAddress } from '@streamr/utils'
+import { Logger, toEthereumAddress, waitForCondition } from '@streamr/utils'
 import { Tracker } from '@streamr/network-tracker'
 import { Stream, StreamPermission, StreamrClient } from 'streamr-client'
-import { fastWallet, waitForCondition } from 'streamr-test-utils'
+import { fastWallet } from 'streamr-test-utils'
 
 import { Broker } from '../../../../src/broker'
 import { createClient, createTestStream, startBroker, startTestTracker } from '../../../utils'
