+++ packages/broker/test/integration/plugins/storage/BatchManager.test.ts
@@ -1,10 +1,9 @@
 import { Client, types as cassandraTypes } from 'cassandra-driver'
-import { waitForCondition } from 'streamr-test-utils'
 import { BatchManager } from '../../../../src/plugins/storage/BatchManager'
 import { BucketId } from '../../../../src/plugins/storage/Bucket'
 import { STREAMR_DOCKER_DEV_HOST } from '../../../utils'
 import { MessageID, StreamMessage, toStreamID } from "streamr-client-protocol"
-import { toEthereumAddress } from '@streamr/utils'
+import { toEthereumAddress, waitForCondition } from '@streamr/utils'
 const { TimeUuid } = cassandraTypes
 
 const contactPoints = [STREAMR_DOCKER_DEV_HOST]
