+++ packages/broker/test/integration/plugins/storage/cassanda-queries.test.ts
@@ -1,6 +1,6 @@
 import { Client } from 'cassandra-driver'
-import { waitForCondition, waitForStreamToEnd } from 'streamr-test-utils'
-import { toEthereumAddress, waitForEvent } from '@streamr/utils'
+import { waitForStreamToEnd } from 'streamr-test-utils'
+import { toEthereumAddress, waitForEvent, waitForCondition } from '@streamr/utils'
 import { Readable, PassThrough } from 'stream'
 import { Storage } from '../../../../src/plugins/storage/Storage'
 import { startCassandraStorage } from '../../../../src/plugins/storage/Storage'
