+++ packages/broker/test/integration/plugins/subscriber/SubscriberPlugin.test.ts
@@ -2,7 +2,8 @@ import { Tracker } from '@streamr/network-tracker'
 import { createClient, startTestTracker } from '../../../utils'
 import { SubscriberPlugin } from '../../../../src/plugins/subscriber/SubscriberPlugin'
 import StreamrClient from 'streamr-client'
-import { fastWallet, waitForCondition } from 'streamr-test-utils'
+import { fastWallet } from 'streamr-test-utils'
+import { waitForCondition } from '@streamr/utils'
 
 const TRACKER_PORT = 12465
 const wallet = fastWallet()
