+++ packages/client/test/end-to-end/resend.test.ts
@@ -1,8 +1,8 @@
-import { fastPrivateKey, fetchPrivateKeyWithGas, waitForCondition } from 'streamr-test-utils'
+import { fastPrivateKey, fetchPrivateKeyWithGas } from 'streamr-test-utils'
 import { createTestStream } from '../test-utils/utils'
 import { range } from 'lodash'
 import { ConfigTest, DOCKER_DEV_STORAGE_NODE } from '../../src/ConfigTest'
-import { wait } from '@streamr/utils'
+import { wait, waitForCondition } from '@streamr/utils'
 import { StreamrClient } from '../../src/StreamrClient'
 import { StreamPermission } from '../../src/permission'
 import { Stream } from '../../src/Stream'
