+++ packages/client/test/integration/Subscriber2.test.ts
@@ -1,8 +1,8 @@
 import 'reflect-metadata'
 
-import { Defer } from '@streamr/utils'
+import { Defer, waitForCondition } from '@streamr/utils'
 import { StreamMessage } from 'streamr-client-protocol'
-import { fastWallet, waitForCondition } from 'streamr-test-utils'
+import { fastWallet } from 'streamr-test-utils'
 import { StreamPermission } from '../../src/permission'
 import { StreamrClient } from '../../src/StreamrClient'
 import { Subscription } from '../../src/subscribe/Subscription'
