+++ packages/client/test/integration/SubscriberKeyExchange.test.ts
@@ -1,6 +1,6 @@
 import 'reflect-metadata'
 
-import { toEthereumAddress } from '@streamr/utils'
+import { toEthereumAddress, waitForCondition } from '@streamr/utils'
 import { Wallet } from 'ethers'
 import {
     ContentType,
@@ -10,7 +10,7 @@ import {
     StreamPartID,
     StreamPartIDUtils
 } from 'streamr-client-protocol'
-import { fastWallet, waitForCondition } from 'streamr-test-utils'
+import { fastWallet } from 'streamr-test-utils'
 import { GroupKey } from '../../src/encryption/GroupKey'
 import { StreamPermission } from '../../src/permission'
 import { Stream } from '../../src/Stream'
