+++ packages/client/test/integration/update-encryption-key.test.ts
@@ -1,7 +1,6 @@
 import 'reflect-metadata'
 
 import { StreamPartID, StreamPartIDUtils } from 'streamr-client-protocol'
-import { waitForCondition } from 'streamr-test-utils'
 import { DecryptError } from '../../src/encryption/EncryptionUtil'
 import { GroupKey } from '../../src/encryption/GroupKey'
 import { Message } from '../../src/Message'
@@ -9,6 +8,7 @@ import { StreamPermission } from '../../src/permission'
 import { nextValue } from '../../src/utils/iterators'
 import { StreamrClient } from './../../src/StreamrClient'
 import { FakeEnvironment } from './../test-utils/fake/FakeEnvironment'
+import { waitForCondition } from '@streamr/utils'
 
 /*
  * Subscriber has subscribed to a stream, and the publisher updates the encryption key for that stream.
