+++ packages/client/test/unit/MessageStream.test.ts
@@ -1,6 +1,5 @@
-import { toEthereumAddress } from '@streamr/utils'
+import { toEthereumAddress, waitForCondition } from '@streamr/utils'
 import { MessageID, toStreamID } from 'streamr-client-protocol'
-import { waitForCondition } from 'streamr-test-utils'
 import { Authentication } from '../../src/Authentication'
 import { createSignedMessage } from '../../src/publish/MessageFactory'
 import { MessageStream } from '../../src/subscribe/MessageStream'
