+++ packages/client/test/unit/OrderingUtil2.test.ts
@@ -5,8 +5,7 @@ import {
     toStreamID
 } from 'streamr-client-protocol'
 import shuffle from 'array-shuffle'
-import { waitForCondition } from 'streamr-test-utils'
-import { EthereumAddress, toEthereumAddress, wait } from '@streamr/utils'
+import { waitForCondition, EthereumAddress, toEthereumAddress, wait } from '@streamr/utils'
 import OrderingUtil from '../../src/subscribe/ordering/OrderingUtil'
 
 const MESSAGES_PER_PUBLISHER = 1000
