+++ packages/network/test/integration/UnixSocketWsServer.test.ts
@@ -1,7 +1,7 @@
 import { ServerWsEndpoint, startHttpServer } from '../../src/connection/ws/ServerWsEndpoint'
 import { PeerInfo } from '../../src/connection/PeerInfo'
 import WebSocket from 'ws'
-import { waitForCondition } from 'streamr-test-utils'
+import { waitForCondition } from '@streamr/utils'
 
 describe('ServerWsEndpoint', () => {
     let serverWsEndpoint: ServerWsEndpoint | undefined = undefined
