+++ packages/network/test/integration/browser-ws-endpoint.test.ts
@@ -1,7 +1,7 @@
 import { Tracker, startTracker } from '@streamr/network-tracker'
 import WebSocket from 'ws'
-import { waitForCondition, runAndWaitForEvents } from 'streamr-test-utils'
-import { waitForEvent } from '@streamr/utils'
+import { runAndWaitForEvents } from 'streamr-test-utils'
+import { waitForCondition, waitForEvent } from '@streamr/utils'
 
 import { ServerWsEndpoint } from '../../src/connection/ws/ServerWsEndpoint'
 import { PeerInfo } from '../../src/connection/PeerInfo'
