+++ packages/network/test/integration/message-duplication.test.ts
@@ -1,8 +1,7 @@
 import { NetworkNode } from '../../src/logic/NetworkNode'
 import { Tracker, startTracker } from '@streamr/network-tracker'
 import { MessageID, StreamMessage, StreamPartIDUtils, toStreamID } from 'streamr-client-protocol'
-import { waitForCondition } from 'streamr-test-utils'
-import { toEthereumAddress, waitForEvent } from '@streamr/utils'
+import { waitForCondition, toEthereumAddress, waitForEvent } from '@streamr/utils'
 
 import { createNetworkNode } from '../../src/composition'
 import { Event as NodeEvent } from '../../src/logic/Node'
