+++ packages/network/test/integration/message-propagation.test.ts
@@ -7,8 +7,7 @@ import {
     StreamPartIDUtils,
     toStreamID
 } from 'streamr-client-protocol'
-import { waitForCondition } from 'streamr-test-utils'
-import { toEthereumAddress, waitForEvent } from '@streamr/utils'
+import { waitForCondition, toEthereumAddress, waitForEvent } from '@streamr/utils'
 
 import { Event as NodeEvent } from '../../src/logic/Node'
 import { createNetworkNode } from '../../src/composition'
