+++ packages/network/test/unit/InstructionThrottler.test.ts
@@ -1,4 +1,4 @@
-import { waitForCondition } from 'streamr-test-utils'
+import { waitForCondition } from '@streamr/utils'
 import { InstructionMessage, toStreamID } from 'streamr-client-protocol'
 
 import { InstructionThrottler } from '../../src/logic/InstructionThrottler'
