+++ packages/network/test/unit/ServerWsEndpoint.test.ts
@@ -3,7 +3,7 @@
 import { w3cwebsocket } from 'websocket'
 import { PeerInfo } from '../../src/connection/PeerInfo'
 import { ServerWsEndpoint, startHttpServer } from '../../src/connection/ws/ServerWsEndpoint'
-import { waitForCondition } from 'streamr-test-utils'
+import { waitForCondition } from '@streamr/utils'
 import NodeClientWsEndpoint from "../../src/connection/ws/NodeClientWsEndpoint"
 
 // eslint-disable-next-line no-underscore-dangle
