+++ packages/network/test/unit/WebRtcEndpoint.test.ts
@@ -1,8 +1,8 @@
-import { MetricsContext } from '@streamr/utils'
+import { waitForCondition, MetricsContext } from '@streamr/utils'
 import { NodeToTracker } from '../../src/protocol/NodeToTracker'
 import { Tracker, TrackerEvent, startTracker } from '@streamr/network-tracker'
 import { PeerInfo } from '../../src/connection/PeerInfo'
-import { waitForCondition, runAndWaitForEvents } from 'streamr-test-utils'
+import { runAndWaitForEvents } from 'streamr-test-utils'
 import { wait, waitForEvent } from '@streamr/utils'
 import { Event as EndpointEvent } from '../../src/connection/webrtc/IWebRtcEndpoint'
 import { RtcSignaller } from '../../src/logic/RtcSignaller'
