+++ packages/utils/test/Metric.test.ts
@@ -1,50 +1,6 @@
 import { wait } from '../src/wait'
 import { AverageMetric, CountMetric, LevelMetric, MetricsContext, MetricsReport, RateMetric } from '../src/Metric'
-
-// this is inlined version of stream-test-utils as we get this error if we'd import this from the library:
-// "Can't resolve '@streamr/utils' in '/Users/teogeb/workspace/streamr/network-monorepo/packages/test-utils/dist/src'"
-// will be fixed in NET-920
-const waitForCondition = async (
-    conditionFn: () => (boolean | Promise<boolean>),
-    timeout = 5000,
-    retryInterval = 100,
-    onTimeoutContext?: () => string
-): Promise<void> => {
-    // create error beforehand to capture more usable stack
-    const err = new Error(`waitForCondition: timed out before "${conditionFn.toString()}" became true`)
-    return new Promise((resolve, reject) => {
-        let poller: NodeJS.Timeout | undefined = undefined
-        const clearPoller = () => {
-            if (poller !== undefined) {
-                clearInterval(poller)
-            }
-        }
-        const maxTime = Date.now() + timeout
-        const poll = async () => {
-            if (Date.now() < maxTime) {
-                let result
-                try {
-                    result = await conditionFn()
-                } catch (err) {
-                    clearPoller()
-                    reject(err)
-                }
-                if (result) {
-                    clearPoller()
-                    resolve()
-                }
-            } else {
-                clearPoller()
-                if (onTimeoutContext) {
-                    err.message += `\n${onTimeoutContext()}`
-                }
-                reject(err)
-            }
-        }
-        setTimeout(poll, 0)
-        poller = setInterval(poll, retryInterval)
-    })
-}
+import { waitForCondition } from '../src/waitForCondition'
 
 const REPORT_INTERVAL = 100
 const ONE_SECOND = 1000
